etl_04/requirements.txt
sqlalchemy 
cx-Oracle
pandas 
dask
pyarrow
dask[complete]
dask[dataframe]
distributed
fsspec
adlfs
prefect_dask